refactor idf terminal
